Split v3.0 firmware into its own project, drop the shared board HAL
v2.0 hardware is frozen, so the dual-target tree was a constraint with no
remaining benefit: v3.0 needs to diverge freely for sleep modes and power
management, and keeping v2.0 buildable from the same sources only taxes
that work.

software/main-unit now becomes software/v3.0/main-unit, targeting v3.0
only. The board HAL is gone -- no boards/, no Kconfig board choice, no
per-revision sdkconfig, one build directory:

  - boards/v30/pins.h is inlined into definitions.h
  - the DRV8662 driver (SDM sine synthesis + EN sequencing) is inlined
    into buzzer/buzzer.cpp as static drv8662_* functions
  - the PCNT rotary encoder is inlined into button/button.cpp, along with
    the button role table that boards/v30 used to supply
  - board_input_init/board_boot_shortcut become init_encoder and
    read_boot_shortcut, declared in button/button.h
  - adjust_volume_for_frequency is now a pass-through: the amplitude
    floor it applied was a DRV8833 H-bridge quirk with no v3.0 analogue

This replaces the previously tracked software/v3.0/main-unit, which was a
plain copy of the v2.0 sources and still carried the v2.0 pin map (BLANK
on GPIO13, DCSIN on GPIO44, RTC_INT on GPIO46) and the LEDC/DRV8833
buzzer, so it did not match v3.0 hardware. That copy had already been
deleted from the working tree without being committed.

The frozen v2.0 firmware is untouched at software/v2.0/main-unit. The
dual-target HAL tree remains in history at 08645a26.

Builds clean against ESP-IDF v5.4.1 (esp32s3); no new warnings.

+++ software/v3.0/main-unit/main/display/tlc5951/tlc5951.h
@@ -21,7 +21,7 @@ class Tlc5951 {
  public:
   void init(
       uint8_t gssin = DEFAULT_GSSIN_PIN,
-      uint8_t dcsin = DEFAULT_DCSIN_PIN,
+      int dcsin = DEFAULT_DCSIN_PIN,  // signed: <0 means "no DCSIN chain" (v3.0)
       uint8_t sclk = DEFAULT_TLC_SCK_PIN,
       uint8_t xlat = DEFAULT_XLAT_PIN,
       uint8_t blank = DEFAULT_BLANK_PIN,
